refactor: 클래스명 오타 수정
