+++ src/main/java/lotto/domain/Lotto.java
@@ -1,7 +1,7 @@
 package lotto.domain;
 
-import static lotto.enums.ErrorMassage.DUPLICATE_LOTTO_NUMBER;
-import static lotto.enums.ErrorMassage.INVALID_LOTTO_COUNT;
+import static lotto.enums.ErrorMessage.DUPLICATE_LOTTO_NUMBER;
+import static lotto.enums.ErrorMessage.INVALID_LOTTO_COUNT;
 import static lotto.enums.LottoConfig.LOTTO_COUNT;
 
 import java.util.Collections;
